Add gitignore and create basic Qt GUI window
Added a gitignore file to ignore Python files. This includes the entire idea directory which is considered user-specific data that should not vary.

Added helperWindow GUI using PyQt5. This is just an empty QMainWindow for now. Goal is to eventually house many of the active contour functions to experiment with various methods and parameters visually via the helper.