[WIP] Continue testing on snake
It is difficult to tune the snake for some of the examples I am working on. Need some way of interpretting the parameters and knowing how and when they need to be changed.

The external energy image can be viewed but I need to find a way to view the internal energy in the snkae. It would be nice to give the user a visible way of determining which parameter should be adjusted.